RSS-1244: Fix high CPU bug in Java 21+ with HEC Lookup by providing proxy to upstream Netbeans library
The old Lookup library is a fork of the upstream Netbeans library. Since the Netbeans library is now distributed standalone, there's no reason to carry the fork of it anymore.

Provide a compatibility layer which can be swapped in via dependency substitution for libraries which have not been updated so that they will call the new implementation and avoid excessive CPU use on current versions of Java. This is not a 100% complete implementation, and should not be used for compilation. This is purely a runtime substitution so that older code will call the new library until such a time that it can be updated.

Also, remove the old, unnecessary fork of the library. Consumers of the lookup library are advised to switch to the upstream Netbeans library directly.